geneviz-backend/requirements.txt
Django==4.2.15
djangorestframework==3.14.0
django-cors-headers==4.4.0
pandas==2.2.2
psycopg2-binary>=2.9
django-cors-headers>=4.4
